batch: Extract source annotation helpers
Batch display currently renders batch-source views and owns the line
annotation helper used by session tracking and command paths.

That couples session and command code to display rendering even when those
callers only need batch-source coordinates for cached line changes.

This commit moves the annotation helpers into source_annotation and updates
runtime callers to import that narrower module directly.

Follow-up test commits will retarget import expectations and pin the source
annotation boundary.
