+++ src/git_stage_batch/commands/file_scope/discard_file_to_batch.py
@@ -5,7 +5,7 @@ from __future__ import annotations
 from contextlib import ExitStack
 import sys
 
-from ...batch.display import annotate_with_batch_source
+from ...batch.source_annotation import annotate_with_batch_source
 from ...batch.lifecycle import create_batch
 from ...batch.ownership import BatchOwnership
 from ...batch.query import read_batch_metadata
