+++ src/git_stage_batch/commands/file_scope/discard_to_batch.py
@@ -6,7 +6,7 @@ from collections.abc import Sequence
 from contextlib import ExitStack
 from dataclasses import dataclass
 
-from ...batch.display import annotate_with_batch_source
+from ...batch.source_annotation import annotate_with_batch_source
 from ...batch.lifecycle import create_batch
 from ...batch.query import read_batch_metadata
 from ...batch.source_refresh import acquire_batch_ownership_update_for_selection
