+++ src/git_stage_batch/data/selected_change/hunk_recalculation.py
@@ -5,7 +5,7 @@ from __future__ import annotations
 import subprocess
 from enum import Enum
 
-from ...batch.display import annotate_with_batch_source
+from ...batch.source_annotation import annotate_with_batch_source
 from ...core.diff_parser import (
     acquire_unified_diff,
     build_line_changes_from_patch_lines,
